Do some more work on the dummy simulator.
There are obviously a lot of errors and missing things still
to be resolved but at least this gives an idea of where
we may want to go.

It expects instructions to be a list of instruction lines
(this is not properly implemented yet). This gives
a lot of clutter but generating things to work on a list
of opcodes makes the specification VERY complicated.